+++ plug-ins/pagecurl/pagecurl.c
@@ -891,7 +891,7 @@ get_gradient_samples (GimpDrawable *drawable,
 {
   GimpGradient *gradient;
 
-  gint     n_d_samples;
+  gsize    n_d_samples;
   gdouble *d_samples = NULL;
 
   gradient = gimp_context_get_gradient ();
